focus-tracker/app/api/dependencies.py
from fastapi import Depends, HTTPException
from sqlalchemy.orm import Session
from app.core.database import get_db
from app.core.settings import Settings, get_settings

def get_current_user(db: Session = Depends(get_db), settings: Settings = Depends(get_settings)):
    # Logic to retrieve the current user from the database
    pass

def verify_goal(goal_id: int, db: Session = Depends(get_db)):
    # Logic to verify if the user's goal is valid
    pass

def check_focus_status(user_id: int, db: Session = Depends(get_db)):
    # Logic to check the user's focus status based on the analysis results
    pass